Implement and test Project and ToDoItem classes.
I implemented the Project and ToDoItem classes based on my class diagrams that I developed during the design stage.
These objects made heavy use of getters and validated setters, as well as helper functions.
Now that I'm thinking about it, it may have been much cleaner to create these objects as factory functions to encapsulate
the helper functionalities of the classes.

Additionally, I installed the 'date-fns' package, extracting the format, isValid, startOfTomorrow, compareAsc, and compareDesc functions from the module.
This package made it easy to work with dates in terms of formatting, comparing, and validating.